+++ packages/did-ui-react/src/components/index.ts
@@ -97,5 +97,6 @@ export * from './context/PortkeySendProvider';
 export * from './SignStep/types';
 
 export * from './types';
+export * from './config-provider/utils';
 
 export * from './utils/operation.util';